A bit of cleanup/reorg
