--- src/lib/claims.ts
@@ -1,20 +1,0 @@
-import { Infer, v } from "convex/values";
-
-/**
- * Shared identity-claims contract between the *provider* components that
- * authenticate users and the *core* component.
- *
- * After a provider authenticates a user it produces this plain payload; the app
- * forwards it to the core's `signIn`, which turns it into a session. Providers
- * never call the core directly, they only know how to produce these claims.
- */
-export const vAuthClaims = v.object({
-  /** Provider name, e.g. "password". */
-  provider: v.string(),
-  /** Stable, provider-scoped account identifier (e.g. a username). */
-  providerAccountId: v.string(),
-  /** Arbitrary profile data the provider learned about the user. */
-  profile: v.any(),
-});
-
-export type AuthClaims = Infer<typeof vAuthClaims>;
